Refactor: Move BitbakeProjectScannerClient to server side
Now the BitbakeProjectScanner resides on the client. We keep exchanging
the scan results to the server.
